Minor changes for error handling
Updated functions.py to include error handling when loading the
dictionary.  There were non-standard characters in the dictionary
causing problems on load.  The new code will print the row numbers in
the dictionary where these errors occur so the user can go and
update/fix the dictionary file.

Also renamed the dictionary file and removed a bunch of rows with the
above mentioned issues.